# BackEnd/tshirt-store-api/week-2/openapi.yaml
openapi: 3.0.3
info:
  title: T-Shirt Store API
  description: |
    REST API for the T-Shirt Store capstone project.
    Supports three roles: Manager, Client, and Delivery Person.
  version: 1.0.0
  contact:
    name: Adrian
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ─── Reusable parameters ────────────────────────────────────────
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    ProductIdParam:
      name: productId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    ProductVariantIdParam:
      name: productVariantId
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Graphic Tees
        slug:
          type: string
          example: graphic-tees
        description:
          type: string

    # ── Address ──
...
    ProductSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
          example: Mountain Sunset Tee
        slug:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [active, disabled]
        category:
          $ref: "#/components/schemas/Category"
        primaryImage:
          type: string
          format: uri
          nullable: true
        likesCount:
          type: integer
        isLiked:
          type: boolean
          description: Whether the current user has liked this product (false if not logged in)
        minPrice:
          type: number
          format: float
          description: Lowest variant price for this product
        createdAt:
          type: string
          format: date-time
    ProductDetail:
      allOf:
        - $ref: "#/components/schemas/ProductSummary"
        - type: object
          properties:
            images:
              type: array
              items:
                $ref: "#/components/schemas/ProductImage"
            variants:
              type: array
              items:
                $ref: "#/components/schemas/ProductVariant"
    CreateProductRequest:
      type: object
      required: [name, description, categoryId]
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
          format: int64
    UpdateProductRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        categoryId:
          type: integer
          format: int64
        status:
          type: string
          enum: [active, disabled]
    ProductImage:
      type: object
      properties:
        id:
          type: integer
          format: int64
        publicUrl:
          type: string
          format: uri
        altText:
          type: string
        sortOrder:
          type: integer
        isPrimary:
          type: boolean

    # ── Product Variant ──
    ProductVariant:
      type: object
      properties:
        id:
          type: integer
          format: int64
        productId:
          type: integer
          format: int64
        sizeId:
          type: integer
          format: int64
        colorId:
          type: integer
          format: int64
        sku:
          type: string
          example: MST-BLU-L
        price:
          type: number
          format: float
          example: 29.99
        stock:
          type: integer
          example: 50
        isActive:
          type: boolean
        size:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
              example: L
        color:
          type: object
          properties:
            id:
              type: integer
              format: int64
            name:
              type: string
              example: Blue
            hexCode:
              type: string
              nullable: true
              example: "#0000FF"
    CreateProductVariantRequest:
      type: object
      required: [sizeId, colorId, sku, price, stock]
      properties:
        sizeId:
          type: integer
          format: int64
        colorId:
          type: integer
          format: int64
        sku:
          type: string
        price:
          type: number
          format: float
          minimum: 0
        stock:
          type: integer
          minimum: 0
    UpdateProductVariantRequest:
      type: object
      properties:
        price:
          type: number
          format: float
          minimum: 0
        stock:
          type: integer
          minimum: 0
        isActive:
          type: boolean

    # ── Cart ──
...
    PromoCode:
      type: object
      properties:
        id:
          type: integer
          format: int64
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed_amount]
        discountValue:
          type: number
          format: float
        startsAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
        usageLimit:
          type: integer
        usageCount:
          type: integer
        minimumPurchaseAmount:
          type: number
          format: float
          nullable: true
        isActive:
          type: boolean
    CreatePromoCodeRequest:
      type: object
      required: [code, discountType, discountValue, expiresAt, usageLimit]
      properties:
        code:
          type: string
        discountType:
          type: string
          enum: [percentage, fixed_amount]
        discountValue:
          type: number
          format: float
          minimum: 0.01
        startsAt:
          type: string
          format: date-time
          nullable: true
        expiresAt:
          type: string
          format: date-time
        usageLimit:
          type: integer
          minimum: 1
        minimumPurchaseAmount:
          type: number
          format: float
          nullable: true

    # ── Pagination wrapper ──
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        meta:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            totalItems:
              type: integer
            totalPages:
              type: integer

    # ── Error ──
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: integer
        message:
          type: string
        error:
          type: string
      example:
        statusCode: 400
        message: Validation failed
        error: Bad Request

# ─── Paths ──────────────────────────────────────────────────────────
paths:
...
  /products:
    get:
      tags: [Products]
      summary: List products with pagination (public)
      operationId: listProducts
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: categoryId
          in: query
          schema:
            type: integer
            format: int64
          description: Filter by category
        - name: search
          in: query
          schema:
            type: string
          description: Search by product name
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/ProductSummary"

    post:
      tags: [Products]
      summary: Create a product (manager only)
      operationId: createProduct
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductRequest"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized (not a manager)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{productId}:
    get:
      tags: [Products]
      summary: Get product details (public)
      operationId: getProduct
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "200":
          description: Product detail with variants and images
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      tags: [Products]
      summary: Update a product (manager only)
      operationId: updateProduct
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductRequest"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      tags: [Products]
      summary: Soft-delete a product (manager only)
      operationId: deleteProduct
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "204":
          description: Product deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{productId}/images:
    post:
      tags: [Products]
      summary: Upload an image for a product (manager only)
      operationId: uploadProductImage
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                altText:
                  type: string
                isPrimary:
                  type: boolean
                  default: false
      responses:
        "201":
          description: Image uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImage"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{productId}/images/{imageId}:
    delete:
      tags: [Products]
      summary: Delete a product image (manager only)
      operationId: deleteProductImage
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Image deleted
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Product Variants ──────────────────────────────────────────────────────────
  /products/{productId}/variants:
    get:
      tags: [Product Variants]
      summary: List product variants for a product (public)
      operationId: listProductVariants
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      responses:
        "200":
          description: List of product variants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProductVariant"
        "404":
          description: Product not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Product Variants]
      summary: Create a product variant for a product (manager only)
      operationId: createProductVariant
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductVariantRequest"
      responses:
        "201":
          description: Product variant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
        "400":
          description: Validation error or duplicate size+color combination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /products/{productId}/variants/{productVariantId}:
    patch:
      tags: [Product Variants]
      summary: Update a product variant (manager only)
      operationId: updateProductVariant
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/ProductIdParam"
        - $ref: "#/components/parameters/ProductVariantIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductVariantRequest"
      responses:
        "200":
          description: Product variant updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductVariant"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Product variant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Likes ─────────────────────────────────────────────────────────
...
  /promo-codes:
    get:
      tags: [Promo Codes]
      summary: List promo codes (manager only)
      operationId: listPromoCodes
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
      responses:
        "200":
          description: Paginated promo code list
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PromoCode"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    post:
      tags: [Promo Codes]
      summary: Create a promo code (manager only)
      operationId: createPromoCode
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromoCodeRequest"
      responses:
        "201":
          description: Promo code created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /promo-codes/{promoCodeId}:
    patch:
      tags: [Promo Codes]
      summary: Update or disable a promo code (manager only)
      operationId: updatePromoCode
      security:
        - BearerAuth: []
      parameters:
        - name: promoCodeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                isActive:
                  type: boolean
                usageLimit:
                  type: integer
                  minimum: 1
                expiresAt:
                  type: string
                  format: date-time
      responses:
        "200":
          description: Promo code updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoCode"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Promo code not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Delivery (optional) ──────────────────────────────────────────
...